requirements.txt
# Web Framework & API
Django~=5.2.0
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.5.0
django-cors-headers~=4.9.0
django-filter~=25.2

# Database & Environment
psycopg[binary]~=3.2.0
python-dotenv~=1.0.0

# External Services (Email & Media)
cloudinary~=1.41.0
django-anymail[sendgrid]~=14.0

# Production & Deployment
gunicorn~=23.0.0
whitenoise~=6.8.0

# Testing & Linting (Development)
pytest~=8.3.0
pytest-django~=4.9.0
black~=25.1.0
ruff~=0.9.0